fixes, optimizations, refactoring

+++ backend/sessionTypes.go
@@ -15,15 +15,15 @@ type PLayerMovements struct {
 }
 
 type PlayerCombat struct {
-	HP              uint8
-	Damage          uint8
-	Kills           uint8
+	HP              int16
+	Damage          int16
+	Kills           uint16
 	WeaponType      uint8
 	WeaponWidth     uint8
 	WeaponRange     uint16
-	LastHP          uint8
-	LastDamage      uint8
-	LastKills       uint8
+	LastHP          int16
+	LastDamage      int16
+	LastKills       uint16
 	LastWeaponType  uint8
 	LastWeaponWidth uint8
 	LastWeaponRange uint16
@@ -54,12 +54,12 @@ type Projectile struct {
 	X, Y           float32
 	VX, VY         float32
 	OwnerId        string
-	Damage         uint8
+	Damage         int16
 	Radius         float32
 	Width          uint8
 	Range          uint16
 	ProjectileType uint8
-	ProjectileId   uint32
+	ProjectileId   uint16
 	LifeTime       float32
 	SpawnTime      time.Time
 }
